1. Не работало переназначение клавиш записи макроса.

+++ far/config.cpp
@@ -1445,16 +1445,16 @@ void Options::Load()
 	HelpTabSize = DefaultTabSize; // пока жестко пропишем...
 
 
-	if ((Macro.KeyMacroCtrlDot=KeyNameToKey(Macro.strKeyMacroCtrlDot)) == -1)
+	if ((Macro.KeyMacroCtrlDot=KeyNameToKey(Macro.strKeyMacroCtrlDot)) == static_cast<DWORD>(-1))
 		Macro.KeyMacroCtrlDot=KEY_CTRLDOT;
 
-	if ((Macro.KeyMacroRCtrlDot=KeyNameToKey(Macro.strKeyMacroRCtrlDot)) == -1)
+	if ((Macro.KeyMacroRCtrlDot=KeyNameToKey(Macro.strKeyMacroRCtrlDot)) == static_cast<DWORD>(-1))
 		Macro.KeyMacroRCtrlDot=KEY_RCTRLDOT;
 
-	if ((Macro.KeyMacroCtrlShiftDot=KeyNameToKey(Macro.strKeyMacroCtrlShiftDot)) == -1)
+	if ((Macro.KeyMacroCtrlShiftDot=KeyNameToKey(Macro.strKeyMacroCtrlShiftDot)) == static_cast<DWORD>(-1))
 		Macro.KeyMacroCtrlShiftDot=KEY_CTRLSHIFTDOT;
 
-	if ((Macro.KeyMacroRCtrlShiftDot=KeyNameToKey(Macro.strKeyMacroRCtrlShiftDot)) == -1)
+	if ((Macro.KeyMacroRCtrlShiftDot=KeyNameToKey(Macro.strKeyMacroRCtrlShiftDot)) == static_cast<DWORD>(-1))
 		Macro.KeyMacroRCtrlShiftDot=KEY_RCTRL|KEY_SHIFT|KEY_DOT;
 
 	EdOpt.strWordDiv = strWordDiv;
